requirements.txt
beautifulsoup4==4.8.1
dj-database-url==0.5.0
Django==2.2.7
django-bootstrap4==1.0.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
Markdown==3.1.1
psycopg2==2.8.4
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
whitenoise==5.0
